refactor(core): WS5 — relocate the acceptance subsystem into the BoringStack adapter
The WS3 documented follow-up. The acceptance/e2e-generation subsystem (acceptance-spec,
acceptance.types, acceptance-outcome, acceptance-steer) was still in core `loop/acceptance/`
but is WEB/SaaS-shaped (nav/shows/screens, testids, negatives-from-rules) and — confirmed by a
full importer scan — consumed EXCLUSIVELY by the BoringStack adapter (build.ts, db-oracle,
gate-stages, and the e2e-generator/e2e-runner/testid-contract already under
loop/boringstack/acceptance/). ZERO core-loop files import it.

Move the 4 files into loop/boringstack/acceptance/ (beside their only consumers) and rewrite the
import paths (moved acceptance-spec now reaches core planning via ../../planning; the 6 boringstack
importers and 12 test files updated). Core now keeps only the generic plan spine
(IProductPlan<TUi>/ISlice<TUi> in planning/). No behaviour change — pure relocation; the acceptance
+ boringstack unit suites (206 tests) and the WS4 boundary test all pass, and `bun run validate`
is green, so the core↔adapter lint boundary holds after moving a whole subsystem across it.
