+++ packages/core/src/loop/boringstack/acceptance/acceptance-outcome.ts
